+++ docs/Changelog.md
@@ -13,6 +13,8 @@ sending a mail to <afl-users+subscribe@googlegroups.com>.
 
   - use -march=native if available
   - most tools now check for mistyped environment variables
+  - the memory safety checks are now disabled for a little more speed during
+    fuzzing (only affects creating queue entries), can be toggled in config.h
   - afl-fuzz:
      - now prints the real python version support compiled in
      - set stronger performance compile options and little tweaks